chore: Delete lib.rs and move it's contents to utils and main.
