+++ core/src/builder/sighash.rs
@@ -198,6 +198,7 @@ pub fn create_nofn_sighash_stream(
 /// and Disprove TX for each sequential_collateral_tx and kickoff_utxo. It yields the sighashes for these tx's for the input that has operators burn connector.
 /// Possible future optimization: Each verifier already generates some of these TX's in create_operator_sighash_stream()
 /// It is possible to for verifiers somehow return the required sighashes for operator signatures there too. But operators only needs to use sighashes included in this function.
+#[allow(clippy::too_many_arguments)]
 pub fn create_operator_sighash_stream(
     db: Database,
     operator_idx: usize,
